move into o.a.k.q
