+++ gradle/spotbugs-exclude.xml
@@ -360,7 +360,7 @@ For a detailed description of spotbugs bug categories, see https://spotbugs.read
         <!-- Suppress a warning about ignoring the return value of await.
              This is done intentionally because we use other clues to determine
              if the wait was cut short.  -->
-        <Package name="org.apache.kafka.common.utils"/>
+        <Package name="org.apache.kafka.queue"/>
         <Source name="KafkaEventQueue.java"/>
         <Bug pattern="RV_RETURN_VALUE_IGNORED,RV_RETURN_VALUE_IGNORED_BAD_PRACTICE"/>
     </Match>
